+++ changelog.md
@@ -9,6 +9,7 @@
   packages.
 * Add the attribute `exe` to derivations in `build.*.executables`, containing the full path to the executable in the
   store.
+* WIP: Add the option `nixpkgs` for configuring different nixpkgs versions.
 
 # 0.9.0